# submission/api/openapi.yaml
openapi: 3.1.0
info:
  title: LabelHub Core API
  version: "1.0.0"
  summary: HTTP edge of the LabelHub data-annotation platform.
  description: >
    Portable OpenAPI description of LabelHub's **HTTP route handlers**
    (`src/app/api/**`). These thin external edges authenticate, then delegate to
    the library layer. They cover the **machine-to-machine** surfaces:
    workspace-API-key ingest/read endpoints, admin (Supabase session) export
    endpoints, public operational endpoints, and the LLM gateway proxy.


    LabelHub deliberately keeps **internal mutations off HTTP** — the browser UI
    creates tasks, submits/reviews annotations, and runs payouts through
    **Next.js Server Actions** (`'use server'` functions), NOT these routes. That
    RPC surface is not expressible in OpenAPI; it is documented in the companion
    `API.md` ("Server-Action RPC surface"). This file documents only what is
    reachable over HTTP.


    ## Authentication models

    Three distinct models are in play (see `securitySchemes`):

    1. **Workspace API key** — machine-to-machine. A `lh_ws_*` (or rate-limited
       public `lh_demo_*`) bearer token. Only the SHA-256 hash is stored; the
       workspace is *inferred from the key*, so these endpoints take no
       `workspaceId`. Accepted on `Authorization: Bearer <key>`, bare
       `Authorization: <key>`, `x-api-key: <key>`, or `x-labelhub-api-key: <key>`.
    2. **Supabase user session** — cookie session from `@supabase/ssr`. Used by
...

...
servers:
...
paths:
...
  /api/llm-assist:
    post:
      tags: [AI assist]
      summary: Per-field / per-topic LLM assist
      description: >
        Signed-in labeler assist. Runs Claude over the field/topic prompt +
        current form context and returns plain text. Per-user rate limit
        (10/min) + daily AI quota. Body capped at 32 KB.
      operationId: llmAssist
      security:
        - supabaseSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [promptTemplate]
              properties:
                scope: { type: string, enum: [field, topic], default: field }
                promptTemplate: { type: string, minLength: 1, maxLength: 8000 }
                context: { type: object, additionalProperties: true }
                tier: { type: string, enum: [fast, default, premium], default: fast }
                itemData: {}
                schemaSummary:
                  type: array
                  maxItems: 80
                  items:
                    type: object
                    properties:
                      id: { type: string, maxLength: 120 }
                      label: { type: string, maxLength: 200 }
                      kind: { type: string, maxLength: 80 }
                      required: { type: boolean }
      responses:
        "200":
          description: Generated text + token usage.
          content:
            application/json:
              schema:
                type: object
                properties:
                  text: { type: string }
                  usage:
                    type: object
                    properties:
                      model: { type: string }
                      inputTokens: { type: integer }
                      outputTokens: { type: integer }
        "400":
          description: Invalid JSON or request.
        "401":
          description: Sign-in required.
        "413":
          description: Body exceeds 32 KB.
        "429":
          description: Per-user rate limit or daily quota exceeded.
        "502":
          description: Upstream AI call failed.
    options:
      tags: [AI assist]
      summary: CORS preflight
      operationId: llmAssistPreflight
      responses:
        "204":
          description: "No content (Allow: POST)."

  # ───────────────────────── Workspace (session auth) ────────────────────────
...
components:
  securitySchemes:
    workspaceApiKey:
      type: http
      scheme: bearer
      description: >
        Workspace API key (`lh_ws_*`, or rate-limited public `lh_demo_*`).
        Accepted on `Authorization: Bearer <key>`, bare `Authorization: <key>`,
        `x-api-key: <key>`, or `x-labelhub-api-key: <key>`. The workspace is
        inferred from the key.
    supabaseSession:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: >
        Supabase cookie session (`@supabase/ssr`). The route resolves the target
        workspace, then enforces a workspace-scoped guard
        (`requireWorkspaceAdmin` / `requireWorkspaceMember`). Cookie name varies
        by Supabase project ref.
    adminToken:
      type: http
      scheme: bearer
      description: >
        Out-of-band ops secret from the `ADMIN_DIAG_TOKEN` env. Send via
        `Authorization: Bearer <token>` or `x-admin-token`; `?token=` query is a
        soft-deprecated fallback.
...